+++ zl-ldm-manha/aula11/agenda-backend/src/main/kotlin/edu/curso/agenda/repository/ContatoRepositoryInMemory.kt
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component
 
 @Component
 class ContatoRepositoryInMemory {
-    var indice : Int = 3
+    var indice : Long = 3
     val listaContatos : MutableList<Contato> = mutableListOf(
         Contato(id=1, nome="João Silva", telefone="(11) 1111", email="joao@teste.com"),
         Contato(id=2, nome="Maria Silva", telefone="(11) 2222", email="maria@teste.com"),
@@ -21,13 +21,13 @@ class ContatoRepositoryInMemory {
         indice++
     }
 
-    fun remover(id : Int) {
+    fun remover(id : Long) {
         val listaFiltrada = listaContatos.filter{
             contato -> contato.id == id}
         listaContatos.removeAll( listaFiltrada )
     }
 
-    fun atualizar(id : Int, contato: Contato) {
+    fun atualizar(id : Long, contato: Contato) {
         for ( indice in 0..< listaContatos.size) {
             val c = listaContatos[indice]
             if (c.id == id) {
